backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.9
pydantic==2.11.4
python-dateutil==2.9.0.post0
pytest==8.3.5
